+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Added
 
 - Add licensing information to the about page.
+- Show delivery comment in STAC browser and on admin page.
 
 ### Changed